NIFI-13579 Incorporated timezone logic in ObjectTimestampFieldConverter.
